add tests for Cart and CreditCardPayment model

+++ src/main/java/assignment/thereadingroom/model/PaymentMethod.java
@@ -1,5 +1,5 @@
 package assignment.thereadingroom.model;
 
 public abstract class PaymentMethod {
-    public abstract boolean pay(float amount) throws Exception;
+    public abstract boolean pay(double amount) throws Exception;
 }
